Update BFS to include a path walkback and implemented metrics()
BFS now keeps a running dictionary to use when recording which node was added to the frontier by which other node. This let's the code walk back through the dictionary like a linked list until the origin node is found. This creates a path from the destitation to the origin. The path is reversed, ie origin to destination, before being handed off to the metrics function.

The metrics function is similar to what Dr. Bennett had planned with the backtrack function. This function creates the various used by the driver functions to create the graphics. It calculates the IDs of each node in the path, their latitudes, their longitudes, and the total distance of the path.

I'm a little unclear about how to best calculate the distance of the path. For now, I am just summing up the euclidian distance between each point and the next in the path. The results are consistent, but usually considerably low. I am unsure how this affects the usefulness of the knowing the path's total distance.